refactor: seed from single file and run all functions in that main.
