+++ package.json
@@ -9,9 +9,9 @@
     "lint": "next lint",
     "migration:generate": "drizzle-kit generate",
     "migration:apply": "drizzle-kit migrate",
+    "db:seed": "npx tsx ./drizzle/seed/seed.ts",
     "test": "jest",
-    "test:watch": "jest --watch",
-    "db:seed": "npx tsx ./drizzle/seed.ts"
+    "test:watch": "jest --watch"
   },
   "dependencies": {
     "@heroicons/react": "^1.0.6",
